Split Qwen35 engine setup
