+++ crates/runner/src/main.rs
@@ -16,6 +16,7 @@ mod policy;
 mod prefill_engine;
 mod profiling;
 mod qwen35_dflash_engine;
+mod qwen35_engine_setup;
 mod qwen35_kv_trace;
 mod qwen35_prefill;
 mod qwen35_runtime;
